Validate conversion grant amount

+++ ApplyToBecomeInternal/ApplyToBecome.Data/Models/AcademyConversionProject.cs
@@ -38,7 +38,7 @@ namespace ApplyToBecome.Data.Models
 		public string AcademyTypeAndRoute { get; set; }
 		public DateTime? ProposedAcademyOpeningDate { get; set; }
 		public bool? SchoolAndTrustInformationSectionComplete { get; set; }
-		public int ConversionSupportGrantAmount { get; set; }
+		public decimal ConversionSupportGrantAmount { get; set; }
 		public string ConversionSupportGrantChangeReason { get; set; }
 
 		//general info
